requirements.txt
# ============================================
# PYTHON BACKEND DEPENDENCIES
# ============================================

# Web Framework (choose one or both based on your needs)
fastapi==0.115.6
uvicorn[standard]==0.34.0
# flask==3.1.0  # Alternative if using Flask

# Database & ORM
psycopg2-binary==2.9.10  # PostgreSQL adapter
sqlalchemy==2.0.36  # ORM (optional)
alembic==1.14.0  # Database migrations (optional)

# Supabase
supabase==2.10.0
# postgrest-py is installed automatically as a dependency

# Google Gemini API
google-generativeai==0.8.3

# LangChain
langchain==0.3.14
langchain-google-genai==2.0.9
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-experimental==0.3.3  # For pandas agent

# Document Processing & RAG
chromadb==0.5.23  # Vector database
pypdf==5.1.0  # PDF processing
python-docx==1.1.2  # DOCX processing
python-magic==0.4.27  # File type detection
tiktoken==0.8.0  # Token counting
python-multipart==0.0.9  # For file uploads
<<<<<<< HEAD
openpyxl==3.1.5  # Excel file processing
xlrd==2.0.1  # Legacy Excel file support
gspread==6.1.4  # Google Sheets API
oauth2client==4.1.3  # Google OAuth for Sheets
=======
>>>>>>> 0e7bc1e (feat: Add NL2SQL write operations with confirmation modal)

# Image Processing & Vision
Pillow==11.0.0  # Image processing
pytesseract==0.3.13  # OCR fallback (optional)
openai==1.59.3  # OpenAI Vision API

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# HTTP Requests
httpx<0.28,>=0.26
requests==2.32.3

# Authentication & Security
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # Must be <5.0.0 for passlib compatibility
cryptography==44.0.0
authlib==1.3.2  # OAuth support
itsdangerous==2.2.0  # Session security

# CORS
fastapi-cors==0.0.6  # If using FastAPI

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Utilities
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.14.1

# Development
ipython==8.30.0
